refactor(core): move collectToolMocks to @mastra/core/evals
Promote the trajectory→item-tool-mock helper out of the playground so
non-UI callers can derive item mocks from a hydrated trajectory without
depending on @internal/playground. Behavior is unchanged.

- Add collectToolMocks (and its tests) under packages/core/src/evals/
- Re-export from @mastra/core/evals
- Playground dialogs now import from @mastra/core/evals; dropped the
  local ToolCallTrajectoryStep cast in favor of the canonical
  TrajectoryStep type
