+++ packages/core/src/evals/collect-tool-mocks.test.ts
@@ -1,28 +1,27 @@
 import { describe, expect, it } from 'vitest';
-
-import { collectToolMocks } from '../collect-tool-mocks';
-import type { ToolCallTrajectoryStep } from '../collect-tool-mocks';
+import { collectToolMocks } from './collect-tool-mocks';
+import type { TrajectoryStep } from './types';
 
 describe('collectToolMocks', () => {
   it('maps tool_call and mcp_tool_call steps to item tool mocks in order, stripping span-name labels', () => {
-    const steps: ToolCallTrajectoryStep[] = [
+    const steps: TrajectoryStep[] = [
       { name: "tool: 'getWeather'", stepType: 'tool_call', toolArgs: { city: 'Seattle' }, toolResult: { temp: 52 } },
       {
         name: "mcp_tool: 'search' on 'docs-server'",
         stepType: 'mcp_tool_call',
         toolArgs: { q: 'forecast' },
-        toolResult: ['a'],
+        toolResult: { hits: ['a'] },
       },
     ];
 
     expect(collectToolMocks(steps)).toEqual([
       { toolName: 'getWeather', args: { city: 'Seattle' }, output: { temp: 52 } },
-      { toolName: 'search', args: { q: 'forecast' }, output: ['a'] },
+      { toolName: 'search', args: { q: 'forecast' }, output: { hits: ['a'] } },
     ]);
   });
 
   it('falls back to the raw label when it does not match the tool span-name format', () => {
-    const steps: ToolCallTrajectoryStep[] = [
+    const steps: TrajectoryStep[] = [
       { name: 'weatherInfo', stepType: 'tool_call', toolArgs: { city: 'Paris' }, toolResult: { temp: 18 } },
     ];
 
@@ -32,13 +31,13 @@ describe('collectToolMocks', () => {
   });
 
   it('walks non-tool container children depth-first to preserve recorded call order', () => {
-    const steps: ToolCallTrajectoryStep[] = [
-      { name: 'first', stepType: 'tool_call', toolArgs: {}, toolResult: 1 },
+    const steps: TrajectoryStep[] = [
+      { name: 'first', stepType: 'tool_call', toolArgs: {}, toolResult: { v: 1 } },
       {
         name: 'agent',
         stepType: 'agent_run',
         children: [
-          { name: 'nested', stepType: 'tool_call', toolArgs: { x: 1 }, toolResult: 2 },
+          { name: 'nested', stepType: 'tool_call', toolArgs: { x: 1 }, toolResult: { v: 2 } },
           { name: 'deeper', stepType: 'workflow_step', children: [{ name: 'leaf', stepType: 'tool_call' }] },
         ],
       },
@@ -48,7 +47,7 @@ describe('collectToolMocks', () => {
   });
 
   it('collects the sub-agent delegation call but not its internal tool calls', () => {
-    const steps: ToolCallTrajectoryStep[] = [
+    const steps: TrajectoryStep[] = [
       {
         name: "tool: 'agent-balanceAgent'",
         stepType: 'tool_call',
@@ -78,21 +77,21 @@ describe('collectToolMocks', () => {
   });
 
   it('does not set matchArgs for ordinary (non-sub-agent) tool calls', () => {
-    const steps: ToolCallTrajectoryStep[] = [
+    const steps: TrajectoryStep[] = [
       { name: "tool: 'getWeather'", stepType: 'tool_call', toolArgs: { city: 'Seattle' }, toolResult: { temp: 52 } },
     ];
     const [mock] = collectToolMocks(steps);
     expect(mock).toEqual({ toolName: 'getWeather', args: { city: 'Seattle' }, output: { temp: 52 } });
-    expect('matchArgs' in mock).toBe(false);
+    expect(mock && 'matchArgs' in mock).toBe(false);
   });
 
   it('ignores non-tool steps and defaults missing args to an empty object', () => {
-    const steps: ToolCallTrajectoryStep[] = [
+    const steps: TrajectoryStep[] = [
       { name: 'gen', stepType: 'model_generation' },
-      { name: 'noArgs', stepType: 'tool_call', toolResult: 'ok' },
+      { name: 'noArgs', stepType: 'tool_call', toolResult: { ok: true } },
     ];
 
-    expect(collectToolMocks(steps)).toEqual([{ toolName: 'noArgs', args: {}, output: 'ok' }]);
+    expect(collectToolMocks(steps)).toEqual([{ toolName: 'noArgs', args: {}, output: { ok: true } }]);
   });
 
   it('returns an empty array for undefined or empty steps', () => {
