+++ packages/core/src/evals/index.ts
@@ -1,3 +1,4 @@
 export * from './types';
 export * from './base';
 export * from './run';
+export * from './collect-tool-mocks';
