+++ packages/playground/src/domains/observability/components/trace-as-item-dialog.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import { collectToolMocks } from '@mastra/core/evals';
 import type { SpanRecord } from '@mastra/core/storage';
 import { useSpanDetail } from '@mastra/playground-ui';
 import type { SideDialogRootProps } from '@mastra/playground-ui/components/SideDialog';
@@ -7,8 +8,6 @@ import { TextAndIcon, getShortId } from '@mastra/playground-ui/components/Text';
 import { useMastraClient } from '@mastra/react';
 import { useQuery } from '@tanstack/react-query';
 import { EyeIcon } from 'lucide-react';
-import { collectToolMocks } from './collect-tool-mocks';
-import type { ToolCallTrajectoryStep } from './collect-tool-mocks';
 import { SaveAsDatasetItemDialog } from '@/domains/datasets/components/save-as-dataset-item-dialog';
 
 type TraceAsItemDialogProps = {
@@ -85,7 +84,7 @@ export function TraceAsItemDialog({
       : undefined;
 
   // Derive item-level tool mocks from the recorded tool calls in the trajectory
-  const toolMocks = trajectory?.steps ? collectToolMocks(trajectory.steps as ToolCallTrajectoryStep[]) : [];
+  const toolMocks = trajectory?.steps ? collectToolMocks(trajectory.steps) : [];
   const initialToolMocks = toolMocks.length > 0 ? JSON.stringify(toolMocks, null, 2) : undefined;
 
   return (
